fix(parser): fail loudly when DeepDoc is unreachable in Go PDF parsing
The Go PDF parser silently fell back to a layout-less MockDocAnalyzer
when DEEPDOC_URL was unset or the DeepDoc inference service was
unreachable, flattening tables and figures into garbled plain-text
chunks. Resolve the endpoint instead (default http://localhost:9390)
and return an explicit error when no healthy service is reachable.
